Decompose GameScene: extract TowerManager and CreepManager
GameScene: 1376 → 1160 lines (-216). Pure refactor, no behavior change.

TowerManager (282 lines) now owns:
- Tower placement (mobile + normal with path validation)
- Tower selling (refund, grid removal)
- Per-frame tower updates (aura resets, trait updates, gold/damage
  collection, firing, alive time tracking)
- Expired tower cleanup (kamikaze, expired Imps)
- Wave-end tower processing (mobile reset, expiry countdown,
  decay per wave, life gain, leak absorb recharge)
- Brood Mother swarmling spawning
- Effective cost calculation with modifier

CreepManager (96 lines) now owns:
- Creep movement updates
- Leak handling (returns LeakResult with damage count)
- Kill gold processing
- Dead creep cleanup
- Creep proximity search (findCreepNear)

Bridge pattern: GameScene.towers/creeps are getter/setters that
delegate to the manager arrays. All existing code that reads
this.towers or this.creeps transparently uses the manager's data.

GameScene update() now calls:
1. towerMgr.updateTowers(time, delta, creeps)
2. creepMgr.update(delta) → returns leak damage
3. towerMgr.cleanupExpired()
4. towerMgr.onWaveEnd() → returns lives gained
